Add initial version of Pisot recurrence prover
This initial version has some bugs. If the conjectured linear recurrence
has a "weird" characteristic polynomial, then the root finding part
breaks down. (For example, [5, 2, 1, 1, 1, ...] gives rise to the third
degree recurrence [[5, 2, 1], [1, 0, 0]], which is a degenerate case in
the code.)

I'll try to play with Zeilberger's Maple code to see what happens in
these cases. The article only considers very nice characteristic
polynomials, so maybe his code is restricted to them as well.